get all spots

+++ backend/routes/api/spots.js
@@ -154,7 +154,7 @@ router.get('/', async (req, res) => {
 
             if (previewImage) {
                 spot.dataValues.previewImage = previewImage.dataValues.url
-                spot.dataValues.avgRating = parseFloat(Number(avgRating)).toFixed(1);
+                spot.dataValues.avgRating = parseInt(Number(avgRating)).toFixed(1);
             }
         }
